requirements.txt
pinecone
google-generativeai
beautifulsoup4
requests
langchain-text-splitters
tiktoken
python-dotenv
